optimize resource load
