+++ common/src/main/resources/i18n_en_US.properties
@@ -63,7 +63,7 @@ updatePersonInfoSuccess=Update personal information successfully
 value=value
 admin.nav.manage=Navigation
 admin.comment.manage=Comment
-templateUpdateSuccess=Change complete
+templateUpdateSuccess=Update template complete
 admin.website.manage=Website settings
 admin.changePwd=Change password
 admin.tag.manage=Tag
