+++ web/src/main/webapp/admin/change_password.jsp
@@ -4,23 +4,23 @@
     <h3>${_res['admin.changePassword']}</h3>
 </div>
 <div class="row">
-    <div class="col-md-6">
+    <div class="col-md-6 col-xs-12 col-sm-12">
         <form role="form" action="api/admin/changePassword" data-toggle="validator" id="updatePasswordAjax"
               class="form-horizontal"
               method="post">
             <div class="form-group">
-                <label class="col-sm-3 control-label no-padding-right"> ${_res['admin.oldPassword']}</label>
+                <label class="col-md-3 control-label no-padding-right"> ${_res['admin.oldPassword']}</label>
 
-                <div class="col-sm-5">
+                <div class="col-md-5">
                     <input type="password" name="oldPassword" value="" required=""
                            class="form-control col-xs-12 col-sm-6" placeholder="" id="orlPassword">
                 </div>
             </div>
 
             <div class="form-group">
-                <label class="col-sm-3 control-label no-padding-right"> ${_res['admin.newPassword']}</label>
-                <div class="col-sm-5">
-                    <input type="password" name="newPassword" value="" class="form-control col-xs-12 col-sm-6"
+                <label class="col-md-3 control-label no-padding-right"> ${_res['admin.newPassword']}</label>
+                <div class="col-md-5">
+                    <input type="password" name="newPassword" value="" class="form-control col-xs-12 col-md-6"
                            required="" placeholder="" id="newPassword1">
 
                 </div>
@@ -30,7 +30,7 @@
             <div class="form-group">
                 <div class="col-md-offset-3 col-md-9">
                     <button id="updatePassword" type="button" class="btn btn-info">
-                        <i class="fa fa-check bigger-110"></i>
+
                         ${_res['submit']}
                     </button>
