+++ web/src/main/webapp/admin/js/article_jqgrid.js
@@ -55,7 +55,7 @@ jQuery(function ($) {
     }
 
     function viewLog(cellvalue, options, rowObject) {
-        return '<a target="_blank" href="admin/article/preview?id=' + rowObject.id + '"><div id="jEditButton_2" class="ui-pg-div ui-inline-edit" onmouseout="jQuery(this).removeClass(\'ui-state-hover\')" onmouseover="jQuery(this).addClass(\'ui-state-hover\');" style="float: left; cursor: pointer; display: block;" title="" data-original-title="' + lang.view + '  ' + rowObject.title.replace(/<font[^>]*?>[\s\S]*?<\/font>/gi, "") + '"><span class="fa fa-search-plus ui-icon"></span></div></a>'
+        return '<a target="_blank" href="admin/article/preview?id=' + rowObject.id + '"><div id="jEditButton_2" class="ui-pg-div ui-inline-edit" onmouseout="jQuery(this).removeClass(\'ui-state-hover\')" onmouseover="jQuery(this).addClass(\'ui-state-hover\');" style="float: left; cursor: pointer; display: block;" title="" data-original-title="' + lang.view + '  ' + rowObject.title.replace(/<font[^>]*?>[\s\S]*?<\/font>/gi, "") + '"><i class="fa fa-search-plus ui-icon"></i></div></a>'
     }
 
     function renderPrivate(cellvalue, options, rowObject) {
