+++ web/src/main/webapp/admin/login.jsp
@@ -4,10 +4,10 @@
 <html>
 <base href="${basePath}"/>
 <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>${init.webSite.title} - ${_res.login}</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1"/>
     <link rel="shortcut icon" href="${basePath}favicon.ico"/>
     <link rel="stylesheet" href="${basePath}assets/css/pnotify.css"/>
     <link rel="stylesheet" type="text/css" href="${basePath}assets/css/login.css"/>
