+++ web/src/main/webapp/admin/plugin.jsp
@@ -1,3 +1,3 @@
 <%@ page session="false"  pageEncoding="UTF-8" %>
-<iframe src="${basePath}admin/plugins/" scrolling="no" style="border: 0px" width="100%" height="1200px"></iframe>
+<iframe src="${basePath}admin/plugins/" scrolling="no" style="border: 0" width="100%" height="1200px"></iframe>
 ${pageEndTag}
