+++ web/src/main/webapp/admin/upgrade.jsp
@@ -8,14 +8,18 @@
     <h3>
         ${_res['admin.upgrade.manage']}
     </h3>
-    <div class="text-right">
-        <button id="checkUpgrade" type="button" class="btn btn-blank">
-            <i class="fa fa-refresh bigger-110"></i> ${_res.checkUpgrade}
-        </button>
+</div>
+<div class="row" style="margin-bottom: 15px">
+    <div class="col-md-12">
+        <div class="text-right">
+            <button id="checkUpgrade" type="button" class="btn btn-blank">
+                <i class="fa fa-refresh bigger-110"></i> ${_res.checkUpgrade}
+            </button>
+        </div>
     </div>
 </div>
 <div class="row">
-    <div class="col-md-6">
+    <div class="col-md-6 col-xs-12 col-sm-12">
         <form role="form" action="api/admin/upgrade/setting" class="form-horizontal" id="upgradeAjax"
               checkBox="upgradePreview">
             <div class="form-group">
@@ -51,7 +55,8 @@
                 <div class="col-md-4">
 			<span class="col-sm-12">
             <label>
-                <input type="checkbox" class="form-control js-switch" style="display: none;" data-switchery="upgradePreview"
+                <input type="checkbox" class="form-control js-switch" style="display: none;"
+                       data-switchery="upgradePreview"
                        name="upgradePreview"
                        <c:if test="${webs.upgradePreview == 1}">checked="checked"</c:if>>
             </label>
@@ -63,7 +68,7 @@
             <div class="form-group">
                 <div class="col-md-offset-3 col-md-9">
                     <button id="upgrade" type="button" class="btn btn-info">
-                        <i class="fa fa-check bigger-110"></i> ${_res.submit}
+                         ${_res.submit}
                     </button>
                 </div>
             </div>
