+++ web/src/main/webapp/include/templates/default-2018/setting/index.jsp
@@ -105,7 +105,7 @@
     <div class="form-group">
         <div class="col-md-offset-3 col-md-9">
             <button type="button" id="template" class="btn btn-info">
-                <i class="fa fa-check bigger-110"></i> 提交
+                ${_res['submit']}
             </button>
         </div>
     </div>
